[refactor] 반환형과 파라미터 Template에 맞게 변경

+++ src/main/java/com/flab/yousinsa/product/domain/events/SellProductEvent.java
@@ -15,6 +15,6 @@ public class SellProductEvent {
 	private Long soldProductOptionId;
 	private Integer purchaseAmount;
 	private Long purchaseOrderId;
-	private Integer remainedStock;
+	private Long remainedStock;
 	private LocalDateTime eventPublishedTime;
 }
